Data.HashMap.Internal.Debug

Description

WARNING

This module is considered internal.

The Package Versioning Policy does not apply.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Authors importing this module are expected to track development closely.

Description

Debugging utilities for HashMaps.

Documentation

valid :: Hashable k => HashMap k v -> Validity k

data Validity k

Constructors

Invalid (Error k) SubHashPath

Valid

data Error k

An error corresponding to a broken invariant.

See HashMap for the documentation of the invariants.

Constructors

INV1_internal_Empty

INV2_Bitmap_unexpected_1_bits !Bitmap

INV3_bad_BitmapIndexed_size !Int

INV4_bitmap_array_size_mismatch !Bitmap !Int

INV5_BitmapIndexed_invalid_single_subtree

INV6_misplaced_hash !Hash

INV7_key_hash_mismatch k !Hash

INV8_bad_Full_size !Int

INV9_Collision_size !Int

INV10_Collision_duplicate_key k !Hash

type SubHash = Word

A part of a Hash with bitsPerSubkey bits.

data SubHashPath