Data.HashMap.Internal.Unsafe

Description

WARNING

This module is considered internal.

The Package Versioning Policy does not apply.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Authors importing this module are expected to track development closely.

Description

This module exports a workaround for this bug:

http://hackage.haskell.org/trac/ghc/ticket/5916

Please read the comments in ghclibrariesbaseGHCST.lhs to understand what's going on here.

Code that uses this module should be compiled with -fno-full-laziness

Documentation

runST :: (forall s. ST s a) -> a

Return the value computed by a state thread. The forall ensures that the internal state used by the ST computation is inaccessible to the rest of the program.