Data.HashMap.Internal.List

Description

WARNING

This module is considered internal.

The Package Versioning Policy does not apply.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Authors importing this module are expected to track development closely.

Description

Extra list functions

In separate module to aid testing.

Documentation

isPermutationBy :: (a -> b -> Bool) -> [a] -> [b] -> Bool

deleteBy :: (a -> b -> Bool) -> a -> [b] -> Maybe [b]

unorderedCompare :: (a -> b -> Ordering) -> [a] -> [b] -> Ordering