Data.HashSet

Description

A set of hashable values. A set cannot contain duplicate items. A HashSet makes no guarantees as to the order of its elements.

The implementation is based on hash array mapped trie. A HashSet is often faster than other tree-based set types, especially when value comparison is expensive, as in the case of strings.

Many operations have a average-case complexity of O(log n). The implementation uses a large base (i.e. 16) so in practice these operations are constant time.

Documentation

data HashSet a

A set of values. A set cannot contain duplicate values.

Construction

empty :: HashSet a

O(1) Construct an empty set.

singleton :: Hashable a => a -> HashSet a

O(1) Construct a set with a single element.

Combine

union :: (Eq a, Hashable a) => HashSet a -> HashSet a -> HashSet a

O(n+m) Construct a set containing all elements from both sets.

To obtain good performance, the smaller set must be presented as the first argument.

unions :: (Eq a, Hashable a) => [HashSet a] -> HashSet a

Construct a set containing all elements from a list of sets.

Basic interface

null :: HashSet a -> Bool

O(1) Return True if this set is empty, False otherwise.

size :: HashSet a -> Int

O(n) Return the number of elements in this set.

member :: (Eq a, Hashable a) => a -> HashSet a -> Bool

O(log n) Return True if the given value is present in this set, False otherwise.

insert :: (Eq a, Hashable a) => a -> HashSet a -> HashSet a

O(log n) Add the specified value to this set.

delete :: (Eq a, Hashable a) => a -> HashSet a -> HashSet a

O(log n) Remove the specified value from this set if present.

Transformations

map :: (Hashable b, Eq b) => (a -> b) -> HashSet a -> HashSet b

O(n) Transform this set by applying a function to every value. The resulting set may be smaller than the source.

Difference and intersection

difference :: (Eq a, Hashable a) => HashSet a -> HashSet a -> HashSet a

O(n) Difference of two sets. Return elements of the first set not existing in the second.

intersection :: (Eq a, Hashable a) => HashSet a -> HashSet a -> HashSet a

O(n) Intersection of two sets. Return elements present in both the first set and the second.

Folds

foldl' :: (a -> b -> a) -> a -> HashSet b -> a

O(n) Reduce this set by applying a binary operator to all elements, using the given starting value (typically the left-identity of the operator). Each application of the operator is evaluated before before using the result in the next application. This function is strict in the starting value.

foldr :: (b -> a -> a) -> a -> HashSet b -> a

O(n) Reduce this set by applying a binary operator to all elements, using the given starting value (typically the right-identity of the operator).

Filter

filter :: (a -> Bool) -> HashSet a -> HashSet a

O(n) Filter this set by retaining only elements satisfying a predicate.

Conversions

Lists

toList :: HashSet a -> [a]

O(n) Return a list of this set's elements. The list is produced lazily.

fromList :: (Eq a, Hashable a) => [a] -> HashSet a

O(n*min(W, n)) Construct a set from a list of elements.

HashMaps

toMap :: HashSet a -> HashMap a ()

O(1) Convert to the equivalent HashMap.

fromMap :: HashMap a () -> HashSet a

O(1) Convert from the equivalent HashMap.